refactor: rename auth middleware dir
