+++ internal/auth/middleware/global_role_middleware.go
@@ -1,4 +1,4 @@
-package authmiddleware
+package middleware
 
 import (
 	"NYCU-SDC/core-system-backend/internal"
